git: Move git.ListTags onto gitserver client
