+++ internal/gitserver/gitdomain/common.go
@@ -154,3 +154,16 @@ type PersonCount struct {
 func (p *PersonCount) String() string {
 	return fmt.Sprintf("%d %s <%s>", p.Count, p.Name, p.Email)
 }
+
+// A Tag is a VCS tag.
+type Tag struct {
+	Name         string `json:"Name,omitempty"`
+	api.CommitID `json:"CommitID,omitempty"`
+	CreatorDate  time.Time
+}
+
+type Tags []*Tag
+
+func (p Tags) Len() int           { return len(p) }
+func (p Tags) Less(i, j int) bool { return p[i].Name < p[j].Name }
+func (p Tags) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
